# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuantVista API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/backtests": {
      "get": {
        "tags": [
          "backtests"
        ],
        "summary": "List Backtests Endpoint",
        "operationId": "list_backtests_endpoint_api_v1_backtests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_BacktestListItem__"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "backtests"
        ],
        "summary": "Submit Backtest Endpoint",
        "description": "Queue a backtest \u2192 202 ``{id, status:\"queued\"}``. Entitlement is checked AFTER validation.",
        "operationId": "submit_backtest_endpoint_api_v1_backtests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitBacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_str__str__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtests/{backtest_id}": {
      "get": {
        "tags": [
          "backtests"
        ],
        "summary": "Get Backtest Endpoint",
        "operationId": "get_backtest_endpoint_api_v1_backtests__backtest_id__get",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Backtest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_BacktestResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "backtests"
        ],
        "summary": "Delete Backtest Endpoint",
        "description": "Remove one run from the history \u2192 204. RLS-scoped: foreign/unknown ids delete nothing \u2192 404.\nAllowed at any status \u2014 an in-flight job no-ops on the missing row rather than resurrecting it.",
        "operationId": "delete_backtest_endpoint_api_v1_backtests__backtest_id__delete",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Backtest Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BacktestListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "universe": {
            "type": "string",
            "title": "Universe"
          },
          "start": {
            "type": "string",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "title": "End"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "type",
          "universe",
          "start",
          "end",
          "created_at"
        ],
        "title": "BacktestListItem"
      },
      "BacktestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "spec": {
            "additionalProperties": true,
            "type": "object",
            "title": "Spec"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "result_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Ref"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "spec",
          "metrics",
          "result_ref",
          "error",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "BacktestResponse",
        "description": "Poll payload for one backtest."
      },
      "BacktestRules": {
        "properties": {
          "rank_by": {
            "type": "string",
            "enum": [
              "composite",
              "fundamental",
              "momentum",
              "quality",
              "sentiment",
              "risk"
            ],
            "title": "Rank By",
            "default": "composite"
          },
          "top_n": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Top N"
          },
          "rebalance": {
            "type": "string",
            "enum": [
              "weekly",
              "monthly",
              "quarterly"
            ],
            "title": "Rebalance",
            "default": "monthly"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "top_n"
        ],
        "title": "BacktestRules"
      },
      "BacktestSpec": {
        "properties": {
          "type": {
            "type": "string",
            "const": "factor_strategy",
            "title": "Type",
            "default": "factor_strategy"
          },
          "universe": {
            "type": "string",
            "const": "NIFTY200",
            "title": "Universe",
            "default": "NIFTY200"
          },
          "rules": {
            "$ref": "#/components/schemas/BacktestRules"
          },
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date",
            "title": "End"
          },
          "costs_bps": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Costs Bps",
            "default": 0
          },
          "benchmark": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Benchmark",
            "default": "NIFTY200_TRI"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rules",
          "start",
          "end"
        ],
        "title": "BacktestSpec",
        "description": "A factor-strategy backtest specification (universe, rules, range, costs, benchmark)."
      },
...
      "Envelope_BacktestResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BacktestResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Error"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope[BacktestResponse]"
      },
...
      "Envelope_dict_str__str__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Error"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope[dict[str, str]]"
      },
...
      "Envelope_list_BacktestListItem__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BacktestListItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Error"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope[list[BacktestListItem]]"
      },
...
      "Error": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Error"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubmitBacktestRequest": {
        "properties": {
          "spec": {
            "$ref": "#/components/schemas/BacktestSpec"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spec"
        ],
        "title": "SubmitBacktestRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}